Moved Sapientino (old DictSpace) to sapientino_env module (not a wrapper).
