+++ gym_sapientino/sapientino_env.py
@@ -23,10 +23,12 @@
 """Sapientino environment with OpenAI Gym interface."""
 
 import random
+import sys
 from abc import ABC, abstractmethod
 from typing import Optional
 
 from gymnasium import Env, Space
+from gymnasium.spaces import Box, Dict, Discrete, Tuple
 
 from gym_sapientino.core.configurations import SapientinoConfiguration
 from gym_sapientino.core.states import SapientinoState, make_state
@@ -74,7 +76,7 @@ class SapientinoBase(Env, ABC):
             self.render()
         return obs, reward, is_finished, False, {}
 
-    def reset(self, seed=None, options=None):
+    def reset(self, *, seed=None, options=None):
         """Reset the environment."""
         if seed:
             self.rng = random.Random(seed)
@@ -118,3 +120,60 @@ class SapientinoBase(Env, ABC):
         # Restore instance attributes (i.e., filename and lineno).
         self.__dict__.update(state)
         self.viewer = None
+
+
+class Sapientino(SapientinoBase):
+    """A Sapientino environment with a dictionary state space.
+
+    The components of the space are:
+    - Robot x coordinate (Discrete)
+    - Robot y coordinate (Discrete)
+    - The orientation (Discrete)
+    - A boolean to check whether the last action was a beep (Discrete)
+    - The color of the current cell (Discrete)
+    """
+
+    def __init__(
+        self,
+        configuration: Optional[SapientinoConfiguration] = None,
+        *args,
+        **kwargs,
+    ):
+        """Initialize the dictionary space."""
+        super().__init__(configuration=configuration, *args, **kwargs)  # type: ignore
+
+        self._discrete_x_space = Discrete(self.configuration.columns)
+        self._discrete_y_space = Discrete(self.configuration.rows)
+        self._x_space = Box(0.0, self.configuration.columns, shape=[1])
+        self._y_space = Box(0.0, self.configuration.rows, shape=[1])
+        self._velocity_space = lambda m, M: Box(m, M, shape=[1])
+        self._theta_space = lambda n: Discrete(n)
+        self._angle_space = Box(0.0, 360.0 - sys.float_info.epsilon, shape=[1])
+        self._beep_space = Discrete(2)
+        self._color_space = Discrete(self.configuration.nb_colors)
+
+        self.observation_space = Tuple([
+            Dict(
+                {
+                    "discrete_x": self._discrete_x_space,
+                    "discrete_y": self._discrete_y_space,
+                    "x": self._x_space,
+                    "y": self._y_space,
+                    "velocity": self._velocity_space(
+                        self.configuration.agent_configs[i].min_velocity,
+                        self.configuration.agent_configs[i].max_velocity,
+                    ),
+                    "theta": self._theta_space(
+                        self.configuration.agent_configs[i].angle_parts,
+                    ),
+                    "angle": self._angle_space,
+                    "beep": self._beep_space,
+                    "color": self._color_space,
+                }
+            )
+            for i in range(self.configuration.nb_robots)
+        ])
+
+    def observe(self, state: SapientinoState):
+        """Observe the state."""
+        return state.to_dict()
