+++ gym_sapientino/wrappers/__init__.py
@@ -22,4 +22,3 @@
 
 """This package contains wrappers for the Sapientino environment."""
 
-from .dict_space import Sapientino  # noqa
